refactor: DTO를 별도의 패키지로 분리
