+++ src/main/java/roomescape/controller/dto/ReservationRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.dto;
 
 public record ReservationRequestDto(String name, String date, String time) {
